+++ src/codegen/CSharpGenerator.js
@@ -1,30 +1,20 @@
 import CodeGenerator from './CodeGenerator.js';
-import template from './templates/csharp.template.js';
+import template from './templates/template.cs?raw';
 
 export default class CSharpGenerator extends CodeGenerator {
-  static get label() {
-    return 'C#';
-  }
-
-  static get language() {
+  get language() {
     return 'csharp';
   }
 
-  static get outputFile() {
+  get outputFile() {
     return 'Program.cs';
   }
 
-  generate() {
-    return this.renderTemplate(template, {
-      SEARCH_URL: this.catalogHref.replace(/\/?$/, '/search'),
-      FILTERS_JSON_INDENTED: this.buildIndentedJson(),
-    });
+  get indent() {
+    return 4;
   }
 
-  buildIndentedJson() {
-    return JSON.stringify(this.filters, null, 4)
-      .split('\n')
-      .map(line => `    ${line}`)
-      .join('\n');
+  get template() {
+    return template;
   }
 }