+++ src/codegen/JavaGenerator.js
@@ -1,31 +1,21 @@
 import CodeGenerator from './CodeGenerator.js';
-import template from './templates/java.template.js';
+import template from './templates/template.java?raw';
 
 export default class JavaGenerator extends CodeGenerator {
-  static get label() {
-    return 'Java';
-  }
-
-  static get language() {
+  get language() {
     return 'java';
   }
 
-  static get outputFile() {
+  get outputFile() {
     return 'StacSearch.java';
   }
 
-  generate() {
-    return this.renderTemplate(template, {
-      SEARCH_URL: this.catalogHref.replace(/\/?$/, '/search'),
-      BODY_STRING: this.buildBodyString(),
-    });
+  get template() {
+    return template;
   }
 
-  buildBodyString() {
-    if (!this.hasFilters()) {
-      return '"{}"';
-    }
-    const jsonStr = JSON.stringify(this.filters).replace(/"/g, '\\"');
-    return `"${jsonStr}"`;
+  formatFilters(filters) {
+    // Run JSON stringify twice to escape the string for embedding in Java code
+    return JSON.stringify(super.formatFilters(filters));
   }
 }