+++ src/codegen/templates/template.rs
@@ -1,10 +1,10 @@
-const template = `use stac::api::Search;
+use stac::api::Search;
 use serde_json::json;
 use stac_io::api;
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let params = json!({{FILTERS_JSON}});
+    let params = json!({{FILTERS}});
     let search: Search = serde_json::from_value(params)?;
     let max_items = search.limit.and_then(|value| usize::try_from(value).ok());
     let items = api::search("{{CATALOG_URL}}", search, max_items).await?;
@@ -17,6 +17,3 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     Ok(())
 }
-`;
-
-export default template;